LeetCode 635. Design Log Storage System
Because the number of operations is very small, we do not need a complicated structure to store the logs: a simple list will do.

Let's focus on the retrieve function. For each granularity, we should consider all timestamps to be truncated to that granularity. For example, if the granularity is 'Day', we should truncate the timestamp '2017:07:02:08:30:12' to be '2017:07:02'. Now for each log, if the truncated timetuple cur is between start and end, then we should add the id of that log into our answer.# Please enter the commit message for your changes. Lines starting