LeetCode 966. Vowel Spellchecker For each word in the wordlist, get its the lower pattern and devowel pattern,
For each lower pattern, record the first such match to hashmap cap.
For each vowel pattern, record the first such match to hashmap vowel.

For each query,
check if it's in the words set,
check if there is a match in cap,
check if there is a match in vowel,
otherwise return "".